requirements.txt
torch
pillow
torchvision
matplotlib
wget